Generate IronMeta on build
Visual Studio runs IronMeta when you save the .ironmeta file, but if you
modify the file outside of VS, or you make a change that the .ironmeta
file depends on, it won't regenerate, and you can have a bug.

To avoid that problem, run ironmeta.exe as a before-build task, to make
sure the generated file is up to date.

I only did this for ParseFasm, since it's currently the only .ironmeta
we're currently working with. We could build proper MSBuild targets and
tasks for this in the future.